GKE Attempt
This was my attempt to move things to GKE. Everything seemed to work
great, but then I realized GKE doesn't support CNI except for some very
tightly controlled deployments. In addition, GKE is a bit too
opinionated for this project, and it's not fair to mandate that everyone
uses GKE. Going back to DIY kubernetes on top of VMs makes it easier for
others to deploy antidote to other platforms.

I am committing this for posterity but also because there are a lot of
good changes in here. I'll be working backwards from here, removing the
GKE specifics but keeping the rest that actually does improve things.

Signed-off-by: Matt Oswalt <matt@keepingitclassless.net>